Add character mappings and use it for title crawl
The end-of-ROM name field is ASCII and the character mappings
interfere with that, so the mappings are lowercase even though the
text appears as uppercase. The couple overlapping characters (space
and the number 2) in the end-of-rom field are changed to literals.

Allegedly ca65 is supposed to have a `.literal` control macro that
ignores character mapping, but that doesn't appear to work for me.