Fix up a bunch of outgoing references from bank7
Where bank7 refers to the same address in multiple banks, I made it
so that there's a global pointer to that that it uses, and the
target banks use the setpos macro to make sure they're aligned to
the same address.

In the future, this would allow for moving them around safely, with
the assembler giving you an error if you move them in a way that
can't really work.