Massive cleanup of the main game state machine
Sets up somewhat-typed info for the three main memory locations
that affect the game's main state loop:

- $0100: the NMI mode, which decides which NMI routine to do and
also acts as a kind of super-mode splitting the game between title
related stuff and main gameplay stuff.
- $0736: the game mode, which is interpreted slightly differently
in the main game from the title screen. This is split into two
variables with different enums: game_mode and title_mode.
- $076C: the game event trigger, which forcefully moves the game
loop from one mode to another on the next NMI. Like $0736, this
also has different meanings on the title screen from the game
itself, so is also split between game_event and title_event.

All of the above are given enums with their values and wherever
I could I made it so that the enum value was used for setting
or comparing.