Update memory map to not exclude stack
This adds a bunch of new bss memory locations in both the area
that is ostensibly for the stack (0x100-0x1ff, especially at the
top and bottom) and an area I had previously mistakenly excluded
as stack (0x200-0x2ff).

I've left in the nominal stack area but commented out in the
nes.cfg file, in case it's useful for the future.

I am assuming that most of the bss_01FX locations are actually
reaching into the stack directly, and not distinct variables.
Some also might be offset table references to info in tables
that are set up in 0x200 and on.