Reduce overworld lag for the most part
This moves the loop for decoding map tiles during horizontal overworld
movement to bank7 so that it can avoid bank switching for every single
iteration of the inner loop of that logic. Bank switching on mmc1
takes a lot of cycles, and the loop will, I think, iterate for N*M
where N is the number of rows on display and M is the number of
changes of tile type on display.